requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
jupyter
kagglehub
ipykernel
pyyaml
joblib
tqdm